fe: fix #1463: require-fail if constraint of type parameter is type parameter
The constraint of a type parameter cannot itself be a type parameter, so this
causes an error.  But the code was not robust against this and crashed with a
require-failure when checking the compatibility of actual type paramaters
against the constraints.

This patch also does some cleanup:

- Moved code to check actual type parameters against their constraints into
  AbstractType.checkActualTypePars to avoid duplication of this code.

- Removed double recursion when checking constraints: This recursion was not
  needed since AbstractType.checkConstraints is called by a visitor over a
  feature that should visit all types anyway, so the recursive checking of type
  parameters is done implicitly.

- Remove NYI-comments for caching (not needed now that recursion is gone and the
  check is cheap), infinite recursion (same) and code-duplication (is removed).

Note that this patch is based on PR #1464, should be reviewed after that PR has
been merged.