lib, me: Fixes when using type features via type pars, fix #1294, fix #1310
First, probably due to the result type of type parameters being wrong until just
recently, some calls to type features in the base lib were unneccessarily
complex and wrong. These have been fixed as follows:

- (float F).type.sum -> F.sum
- (numeric N).type.sum -> N.sum
- (wrapping_integer W).type.all_bits_set -> W.all_bits_set
- (wrapping_integer W).type.byte_size -> W.byte_size

This fixes idioms 54 and 201..203.  For idiom 204 to work a fix was needed in fz
since the code to determine the result type of a call to a type feature was
incorrect if that type feature was inherited.  This patch fixes the lookup
extracting the result type of an inherited type parameter by the actual type
given in the corresponding inherits call.

Additionally, this patch relaxes from checks and postconditions I ran into while
debugging this.

Next to these fixes, this patch also introduces a new regression test that
contains the code of some of the failing idioms, simplified versions thereof and
some more tricky cases of inherited type parameters.

This PR is based on #1463, so should be reviewd only after that PR was merged.