fe: Explicitly show Types.f_ERROR and Types.t_ERROR in error messages
In principle, Types.f_ERROR and Types.t_ERROR can occur only as a consequence of
previous errors.  So any errors reported that contain these should be ignored
since they are only subsequent errors caused by a previously detected problem.

Nevertheless, sometimes these are shown anyways, so this patch marks them in red
color when f_ERROR is shown and prints `**error**` instead of the very confusing
`universe` in case of t_ERROR.

In a later phase, we could lat `fz` crash with a failed `check` if these are
part of errors printed.