air: Remove ugly special handling for result clazz determination
With the fixes for the outer handling, the complex special handling of outer
clazzes is no longer needed, the result clazz can be determined from the resul
tytpe `t` using `actualClazz()`.

There is one case that is not handled correctly when using `actualClazz`: type
features.  This is why we need a workaround for #1477 and similar cases. The
problem is apparanetly that for a field

  y := n.this.type

The result type of feature `y` is `n.#type n.this.type`, which for an actual
outer clazz `i` for `n.this.type` results in `n.#type i`, while we need `i.#type
i`.  This might need to be fixed in the front end during type inference: `y`
result type should probably be `n.this.type`.