num.matrix: major cleanup and fixes
With this change, `num.matrix` no longer inherits from `numeric`. This
is because it is not possible to implement important things from
`numeric`, such as `type.zero`, without specifying an additional
dimension parameter.

In addition to removing the inheritance from `numeric`, restructure the
code, add more comments, and fix the implementation of `type.equality`
(always returned true until now).