air: Remove clazz inference from middle end, fix #1477
The information required to determine the clazz of a type feature's type was
available, but was not used.  This patch still uses a special handling for type
features, but is very similar to the cases just before for `f.isTypeParameter`
and `Types.get`.

The special handling for `ft.isThisType` turned out to be coverd by
`actualClazz` already, so I removed it.

This PR will enabled some cleanup in package `ast` since `Feature.initialValue`
is no longer needed after the front end phase, will do this in a separate PR.

Also, it might be possible to simplify the method further (e..g, there are 3
calls to `typeClazz`);  I will leave this for later.