fe: Fix two subsequent errors in the example from #1505
First, when inherited features cannot be found, not attempt should be made to
call these as parents of the corresponding type features (otherwise we try to
create and call call Types.f_ERROR's type feature resulting in all sorts of
issues).

Second, the type feature should not redefine the 'THIS_TYPE' type parameter if
the inherits list is empty due to previous errors, so there is nothing to
redefine.