lib/Function: remove `Functions` feature
What can be done using `Functions.compose` is now possible in a nicer
way using the `Unary` type and its `infix ∘` operator. It might be nice
to allow more arbitrary function composition in the future as documented
in the now removed comment, but this is not strictly necessary, at least
for now.