fe: Allow redefinition of feature with type parameters, fix #1387
This does allow the execution of the example given in #1387.  The example is
added to a new test case.

However, the new test case also add three variants that currently do not work
yet.  I will add separate issues for those to get fixed.  The main problem is
types used in the signature that depend on type parameters of the redefined
feature.