fe: propagate `unit` expected type for code in constructors, fix #1066
This avoids errors for loops that may produce a boolean result if these loops
are used as the last statement in a constructor.  The infrastructure to suppress
the result in case `unit` is expeced was already in place, what was missing was
providing this information during type inference on the expected result type of
the code of a constructor.