util/fe/tests: Fix check if two positions refer to same source file fix #1550
Since the lexer and parser inherit from SourceFile we end up with many instances
refering to the same file.  This breaks the equality test using `==` and `!=`.
This patch introduces a new feature `SourceFile.sameAs` that handles clones as
equal.

The enable uncommenting a line of code in tests/visibility that did not compile
before.