tests/atomic: Enable concurrent tests for concur.atomic.compare_and_swap
The original versions of these tests do not work since tail recursion
optimization is not performed here, which it should, but this is a separate
issue that needs to be fixed.

This patch replaces these tests by recursive ones that do not perform a deep
recursion, so do not cause a stack overflow.

For better understanding what is happening, the test now understands an option
`-debug` that prints statistics on the compare_and_swap successes and failures.