fe: Fix type handling when calling constructor on type parameter, fix #1590
The main problem was that in a call `T.s` with a type parameter target `T`
calling a constructor `s` the outer type of the result was not set correctly
resulting in wrong replacements of type parameters in case `s` inherits from a
feature with type parameters.

This PR now uses the type parameter `T` as the outer type for `T.s` instead of
`T`'s constraint.

There is still too much special handling for calls involving type parameters, so
I added NYI-comments to eventually remove these.

This patch requires AbstractFeature.typeFeatureOrigin() to work even for library
features, so LibraryFeature was extended to support this

This PR also addes some helper methods to improve code readability:

 - AbstractType.isThisTypeInTypeFeature
 - Call.targetIsTypeParameter
 - Call.targetType
 - Generic.isThisTypeInTypeFeature
 - Generic.typeFeatureOrigin

tests/reg_issue1294 was extended to include an updated version of the original
example fron #1590.