fe: in fixed features, allow covariant redefinition in type pars, fix #1556
A fixed feature `h.f` that redefines `p.f` that uses type`p.this.type` in its
signature already was allowed to use `h` instead.  This patch relaxes this
condition and permits types depending on `p.this.type` in their type parameters
to use `h` instead.

This patch removes the special handling from issue #737 in the case of `new
Type(t, g, o)` being called from `AbstractType.applyToGenericsAndOuter()`, but I
could not fully remove this in all cases, there remain constructors in `Type`
with a boolean `fixOuterThisType` that should be checked and removed later.