fe:
Allow non-fixed redefinition to covariantly change `parent.this.type` to
`heir.this.type` when used as type parameter (as in `option x.this.type`).

Cleanup in code to check covariant redefinition of `this.type` used in fixed
features defined in type features.  Enhanced this code to allow use of
`this.type` as type parameters as in `option x.this.type`.

Improved error handling to suppress additional errors if types in redefinition
checking contain errors.

Added comments, renamed `ignoreFixedModifier` as `fixed` and pull the code to
check the modifier out of `isLegalCovariantThisType`.