fe: Ignore type constraint in Call.handDownFormalArg, fix #1634
handDownFormalArg changes a formal type used in a declared feature into the
actual type of the call target.  This performs replacing formal generics along
he inheritance chain by the actual generics provided in the inheritance calls.

The Problem was that for a target of generic type, the type constraint was used,
which is wron ghere.