fe: Produce an error in case of a declaration in a lazy expression, #1666
Currently, features declared in an expression used as a lazy value will cause
errors since the declaraions are resolved twice, once before and once after they
were wrapped in the Lazy instance.

This patch is only a workaround that produces an error message, explains what to
do alternatively and what needs to be improved in the Fuzion code to properly
fix this.