fe: Add support for nested lazy values, fix #1665
The problem essentially was that in nested calls, actual arguments of inner
calls were mapped before outer calls, which results in wrong outer features for
the instances of Lazy created.

This patch solves this by

- splitting up wrapInLazyAndThenPropagateExpectedType (what was named the wrong
  way around) into two phases: propagateExpectedType and wrapInLazy

- changing the order for wrap*InLazy to process outer calls before actuals in
  inner calls.

As a bonus, this patch adds support for automatic wrapInLazy for assignments
to fields such as

  i Lazy i32 := f x

A regression test testing three levels of nested lazy arguments and lazy
assignments to fields is also added.