parser, base lib, tests: Type `a.this.type` is now written as `a.this`
Just as `a` can be used as a value or a type, `a.this` now can be used as a
value or as a type using the same syntax.

`a.this.type` is only used to get the type of `a.this` as a value, just like
`a.type` is the value of `a`'s type instance.

In most situations, this avoid the need for `.type`.