fe: Fix check failure in case `x` is not found in `x.this.type`, fix #1711
The special handling for `qualThis` to create either an Expression or a Type has
been removed. `qualThis` now produces a list of strings that is then wrapped
into an intance of `This` or `QualThisType`.

`QualThisType.java` is a new class for parsed types of the form
`x.y.this.type`. They no longer have any outer types set, but contain the
qualified name only that will be used to resolve the feature.

The code in `This.java` to look find `x.y` in an expression `x.y.this` was
enhanced to complain about ambiguity and to become usable for `x.y.this.type` as
well.

Failure of feature lookup in a `x.y.this` expression now results in
`Expr.ERROR_VALUE` such that follow-on errors should be suppressed.

Feature lookup in `Type.resolveFeature` now performs special handling for
`x.y.this.type`, where the lookup has to be done using `This.getThisFeature`.

Now, in case of `x.y.this` or `x.y.this.type` expression errors, a correct list
of available qualified names is shown together with the information if the
problem was an ambiguity or a not-found feature.