fe: Permit all calls with `void` as the target type, fix #1724.
Originally, this was required to suppress subsequent errors when type inference
for formal arguments results in void. But I consider this logical (and maybe
even useful) in the general case: If you have an instance of type void, you can
do anything with it!

Added a test that performs different undefined calls on void.

Also removed no longer needed `Function.loadCalledFeature`, this was used only
for `fun`-calls that are no longer supported.