fe: Support free types, fix #1656
Within function formal argument lists, type parameters can not be added inline.
I call these free types, the basic syntax is `T : Constraint` as in

  a(v T : Any) => ...

which is equivalent to

  a(T type : Any, v T) => ...

The constraint can be dropped, i.e., any unqualified type name without actual
type parameters that is not found will be treated as a free type as in

  a(v T) => ...

Free types can be shared by several arguments

  a(v, w T) => ...
  a(V T, w T) => ...

which means that the types must all be the same. Also, free types can be
anonymous using `_`:

  a(v, w _) => ...

which means that the type cannot be referenced explicitly, but there is only one
type parameter and `v` and `w` have to be the same type.

In contrast, several anonymous free types are distinct, i.e.,

  a(v _, w _) => ...

has two different anonymous type parameters, one for `v` and one for `w`.

Anonymous type parameters and type inference for arguments as in

  a(v, w) => ...

is different, the latter infers exactly one static type for v and w, while the
former create a type polymorphic feature.

Free types may be used as type parameters of  other types, e.g, as a choice

  g(v A | (B : numeric)) => ...

or as sequence

  x(s Sequence T) => ...

Next to supporting free types, this PR adds new errors trying to limit confusion
in case of a typo in a type.  For safety, free types must not mask existing types.

This patch adds positve and negetive tests `free_types` and
`free_types_negative` with a number of examples.