fe: Fix saving free type parameters to module file
The problem was that when adding free type parameters, they were added to
`Feature._arguments`, but not to `Feature.arguments0` (which is the same list,
but using type `AbstractFeature` instead of `Feature`. Since this list was used
when saving the module file, the free type parameters were not saved.

This commit fixes the issue by removing changing the type of `arguments` to
`List<AbstractFeature>` and removing `arguments0`. Unfortunately, this adds the
need for some type casts to `Feature`.