analysis: Adding JVM backend infrastructure
In the abstract interpreter, Pop is no longer a nop but actively pops the value
from the stack and drops it.  A new method `AbstractInterpreter.drop()` gives the backend the chance to do this.

This will be required by the JVM backend where values can have side-effects,
so they cannot be discarded without notice.

Also added parent `ANY` to `AbstractInterpreter.ProcessStatements`.