fuir/analysis: When setting argument fields, do not force to use value instance
This is a special handling that was introduced for the C backend where structs
for ref types always contain an embedded corresponding value type instance, so
accesses to fields are always to the fields of the value type.

This special handling in the AbstractInterpreter is no longer needed for the C
backend and it is in the way for the JVM backend that does not create value
types for every ref type.  SO this patch removes that special handling.