jvm: add support for abortable effects
This is similar to the way the Interpreter impelments abortable effects: An
expception `Abort` is thrown and handled by a suitable installed effect handler.

What was alittle tricky for the JVM backend is to permit the callback within
effect_abortable.  This is done now a little manually via passing an instance to
the clazz and searching for `fzRoutine`. This willt need to be improved for
better performance at some point.

Also added some comemnts and support for cpClass constants in class files.