fe: move constraint must not be choice error into its own method
This separates the incompatible type parameter and the constraint must
not be a choice error messages, since they are fundamentally different
things.

Furthermore, remove the ill-advised suggestion for fixing the error
added in the previous commit, since it does not work, and use the
position of the offending type parameter in the error message, instead
of the position of the offending call.