fe: fix #1943: Using type parameter as outer type
I was unsure when I wrote this code if using a type parameter as an outer type
actually makes sense, but it does as shown in the examples of the regression
test.

The bug that is patch fixes was that the constraint type of the type parameter
was not resolved yet and hence not usable.