parser: disallow line break after `->` in function return type, fix #1945
This is an somewhat ugly special handling but the best solution that I could
find so far to avoid the ambiguity between a feature declaration with a function
result type and a call with a lambda argument.

Now, it is no longer legal to write declarations such as

  f () ->
    i32 := ()->42

one must use either

  f (() ->
     i32  ) := ()->42

or

  f () -> i32 := ()->42

while a call

  f ()->
    x := 42

can not be written as

  f ()->x:=42

This patch adds a positive and a negative regression test for #1945.