be/jvm: determine values of debug and safety intrinsics at build time
The value of the `debug`, `debug_level` and `safety` intrinsics is now
determined by the values of these options passed to the `fz` tool at
build time, rather than hard-coding values in the JVM runtime.

This also fixes problems caused by the DFA and the JVM backend using
different values, in case a non-default value was passed to Fuzion at
the build time.