lib: Remove Any.hash_code
Just like equality, the hash_code function should be a type feature and not be
defined for Any and inherited.

For hash_code to be useful, the following invariant should hold:

  (hash_code a != hash_code b) || (a = b)

So the semantics of hash_code depends on type.equality, so hash_code should be a
type feature as well.