jvm: Do not throw away code for dynamic call with unit type result
For a dynamic call with more than one targets, the resulting code is just an
`invokeinterface` to a method performing the actual call. If, however, this call
returns a unit type value and that value, the AbstractInterpreter currently
throws that code away.

This change makes sure that a unit type value is returned as a RESULT (i.e.,
code) not as a VALUE to the AbstractInterpreter, so it will not be ignored.

For non-unit type values, the AbstractInterpreter already makes sure that the
value is drop()ped properly if it is not used.