be/C: Fix failure in tests/ternary by forcing const arrays to be on the heap.
The problem was that a call of the form

  f =>
    [1,2,3].as_list

will pass a ref to the array to `array.as_list`, which will create a Cons cell
that contains that ref. Since that Cons cell with be returned as the list result
of `f`, we have a dangling reference.

Before array consts, this was solved since the array was instanciated via a call
and the call result was heap cloned explicitly.

In a later step, I would like to see all const arrays stored globally and
allocated only once such that all the cloning would no longer be a problem.

As for heap cloning for call results in the C backend: This is done way too, the
plan is to find a common solution for this in the JVM and C backend, so no need
to worry right now.