be/C: Fix failure in tests/hash_map by higher DFA accuracy for SysArray elements
The problem was that the comparison of `SysArrays` did not take the element
values into account, such that joining the `SysArray`s of `[123]` with `[]`
could result in the `SysArray` for `[]`, whose element values was `null`, i.e.,
`void` resulting in DFA to stop analysis when accessing the elements of `[123]`.

This patch fixes `SysArray.compareTo` to compare `_elements`. Additionally, it
adds some infrastucture that may help in the future to extract the element
values from the data array of a constant array.

This patch also add element values for the result of `fuzion.sys.fileio.mmap`
which was missing and worked only by chance.