jvm: Fix compilation of routines that produce a void result
This fixes the code for stubs for calls to features that do never return (Fuzion
result type `void`), these need a `return` bytecode anyway to keep the JVM
classfile verifier happy.

Additionally, a feature's result field may not exist if the code produces a void
value as in

  f i32 => panic "bah"

so we must not attempt to read the result field in the epilog.

This fixes the failure of tests/local_mutate when using the JVM backend.