ast: remove type inference in case of routine with explicit type and `=>`
For a routine declaration of the form

  r() t => v

The result type was ignored and inferred from `v`.  The reason was that the Impl
was set to RoutineDef, while it should be Routine in this caes.

This patch makes sure that if a type is present in a routine declaration, the
kind will be `Routine`, not `RoutineDef`.