jvm: Add support for labels as branch targets when generating bytecode
This is require for more control flow, in particular for tail recursion where we
need a `goto` to the beginning of a method.

The main change is that ByteCode no longer provides a methods `bypeCode` and
`bc*` that produce a `byte[]`, but instead there are methods `code(bw, ...)`
that write bytecode to a `ByteCodeWriter`.

These `ByteCodeWriters` is an abstract class that will be run three times:
first, to get an upper bound estimate for the bytecode size and offsets of
labels, next to then freeze this upper bound and fix all labels, and finally to
produce the code now that the labels adresses are known.

This infrastructure will also be needed to support bytecode `lookupswitch` and
`tableswitch` that have some alignment requirements.

Unfortunately, this patch requires changes in `Choices.java` since labels can
only be used once, so an `Expr` using a label cannot be used for code generation
repeatedly.  The code for the cases in a choice were, however, reused, which had
to be changed such that every case has its own copy of the code with its own
labels.