jvm: Support tail recursion
For this, we first have to make sure that the Java stack is empty when executing
cases in a choice since otherwise a tail call that performs a `goto start` would
be performed with a non-empty stack resulting in classfile verification to fail.

Then, we can replace a tailcall with re-assigning the argument variables and
just performing a `goto`.

This patch also updated the expected error output for two tests since source
lines have changed.