lib: Add collection of features for function composition, #2046
This is not intended to be useful at this point, it is just a first try to
collect basic composition features that appear in literature and other
languages.

To make this useful, we need to write example code using these and play around
with it. Maybe we need to move these into Unary.fz, maybe add Binary.fz, maybe
make some operators or add operator aliases, maybe add features or operators to
the universe, etc. Maybe we want these in a separate module?