jvm: Fix test failure in tests/outerNormalization
No code was generated for a dynamically bound assignment since the wrong result
clazz was used.

This patch moves the local var `rt` outside an `if` such that the result type is
determined only once.

Also, the argument count for an assignment, which should always be 1, was wrong
for a dynamically bound assignment.