jvm: Support assignments of value types, fix tests/assignments failure
Value types are allocated as instances of a Java class. So far, assignments of
value types was done by assigning references to these instances. This is ok only
if the value is immutable, which is not always the case.

This patch creates new instances on assigments of value types.
Performance-wise, this is catastrophic since really many temporary instances are
allocated on the heap.

To improve the performance, the goal is to avoid class instances for value types
altogether and expand value types into a sequence of fields, i.e., to implement

  point(x, y i32) is
  line(a, b point) is
  l := line (point 3 4) point (5 12)

as follows (in pseudo-Java code):

  int temp1_x = 3
  int temp1_y = 4
  int temp2_x = 5
  int temp2_y = 12
  int temp3_a_x = temp1_x
  int temp3_a_y = temp1_y
  int temp3_b_x = temp2_x
  int temp3_b_y = temp2_y
  int l_a_x = temp3_a_x
  int l_a_y = temp3_a_y
  int l_b_x = temp3_b_x
  int l_b_y = temp3_b_y

i.e., the local var `l` will get expanded into four local variables in Java code.