jvm: add checks to trap accidental modifications of immutable array data
This turns the precondition failure in `test/stdin` into a nicer error saying
`error 1: Attempt to modify immutable array`.

The way this works is that all internal array data that are put into an instance
of an (immutable) array will be marked as frozen using a weak map of all frozen
array data pointers. On modification, it will then be checked that the modified
array is not in that map.

These checks are currently enabled only if CHECKS is set. In case debugging is
disabled, eventually no code should be generated for this.