lib: For intrinsics that write arrays, do this before creating immutable array
In general, the code pattern

  arr := [0,0]
  intrinsic_function arr.internal_array.data
  result := arr

is changed into

  iarr := fuzion.sys.internal_array_init T length
  intrinsic_function iarr.add
  result := iarr.as_array

This is a little clumsy, but in many cases the code should be change anyway once
we can return tuples from intrinsics.

Since this removes the code to intialize the array, the DFA no longer sees any
values assigned to these arrays. So this patch adds explicit handling in the DFA
to initialize these arrays.