jvm: Support `compare_and_set/swap` for general choice types
With this, product and sum types can now be used, they are compared and copied
recursively.

When a choice is compared, all the fields are taken into account, also the
hidden fields. This seems to be ok since also when a choice is copied, all
fields are copied.

This patch also adds new test cases to `tests/atomic` for different choice
types.