ast: For actual type parameters, create all required type features
The problem was that for a type passed as an actual type parameter, the type
parameter of that type itself did not get its type feature created, resulting in
an error when finding all used features in the middle end.

This patch forces the creation of all type features for all features appearing
as outer features or as type parameters of an actual type parameter in a call.