lib: use intrinsic_constructor for instrinsics resulting in bool
The example of #2114 was so minimalistic that during findClazzes, no instances
of `TRUE` or `FALSE` where found to be created, resulting in broken C code and
an error in the JVM backend (since type `bool` is effectively `void` when no
instances of `TRUE` nor `FALSE` exist).

Now, using `instrinc_constructor`, a call to one of these intrinsics ensures
that `TRUE` and `FALSE` exist.