jvm: fix checks to not modify immutable array for `array unit`
The internal array data in the JVM backend for `array unit` is just `null` since
there is no data to be stored.  However, this causes problems when one such
array is frozen and a second one is initialized, since the internal array
pointer is the same (`null`), the runtime check that the second one is not
frozen when it is initialized fired.

This patch makes sure that the internal arrays for `array unit` will never be
frozen such that the checks do not apply.