lib/ast: Add `id` routines to base lib, use `id` to wrap `void` call target
Both version of `id` have a type parameter. One with not value argument produces
the identity fuction, while the one with a value argument itself implement the
identity function as a routine.

These are used in a call `x.y` if `x` results in `void`.  This used to be
replaced by `void x`, but `void` is a choice type now that cannot be called. So,
now, this is replaced by `id void x`.