C backend: Fix access to uninitialized memory in compare_and_swap/set, fix #2121
The problem was that compare-and-swap and compare-and-set operations on atomics
did compare padding bytes as well, which could contain random unintialized data.
The reason is that `memcmp` was used for the comparison.

This patch fixes this by providing an implemenation of `compareValues` similar
to the JVM backend that does a structural comparision of values.

With this patch, the assignment still copy along uninitalized padding bytes or
unused union fields. This, however, is no problem since this data will never be
used to affect the execution of the program, while the comparison did.