lib/jvm: Add support for mmap to JVM backend, cleanup of mmap implementation
This patch first introduced a new abstract feature `container.buffer` that is
the basis of `io.file.open.mapped_buffer` that is no used for `mmap`.

The use of `mapped_buffer` instead of `marray` permits the JVM backand to chose
a different impementation for mmapped memory: These are represented as
`java.nio.ByteBuffer`, while `marray u8` internally use Java's `byte[]`.

All the other backends need to impement new intrinsics for `mapped_buffer`, but
can use the original representation.

This patch removes te effect `io.file.open.mmap` and uses `io.file.open`
instead.  Consequently, the code passed to `io.file.open.mmap` no longer needs
to use `io.file.open.mmap` (no arguments) to access the mapped memory. Instead,
I have added an argument `buf` to the lambda that gives access to the memory.

One remaining issue that affects all code using effects: It is ensured that the
mapped buffer is accessed only while `io.file.open T` is installed, but it is
currently not enforced that a buffer is stored in a mutable variable and
accessed later, when a different instance of `io.file.open T` was installed.
This is curently solved only for `mutate` that uses a unique id for this that is
checked at runtime.