jvm: Fix `Choices.kind(bool)` to `boollike`
In very small tests that use type `bool` but do not use both `TRUE` and `FALSE`,
the choice kind would get set to `unitlike` or even `voidlike` resulting in
errors since bool is otherwise hard-coded as type `Z` in the JVM backend.

This patch hard-wires the choice kind for `bool` to `boollike` to avoid this.

This fixes failures of examples from flang.dev like
`content/design/examples/tya_const.fz`.